backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
google-generativeai
slowapi
huggingface_hub[inference]
pillow
